+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nostr-wot/ui
 
+## 0.5.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @nostr-wot/signers@0.4.0
+
 ## 0.5.0
 
 ### Minor Changes